✨ feat(config.py): add configuration file for AI services integration to manage API keys and model settings
🧹 chore(streamlit_app.py): clean up imports and remove unused configurations to enhance code readability and maintainability

🗑️ chore(streamlit_app.py): remove unused functions to clean up the codebase and improve maintainability

🔧 fix(streamlit_app.py): update download function to include mode parameter for improved functionality

✨ feat(summary.py, transcription.py): add new summary and transcription modules to handle audio processing and summarization
🔧 chore(summary.py, transcription.py): implement caching and error handling for improved performance and reliability in audio transcription and summarization processes

✨ feat(translate.py, tts.py, utils.py): add translation and text-to-speech functionality to the application

- Introduce `translate.py` for translating text using a model with caching and error handling.
- Implement `tts.py` for generating audio from text using OpenAI and ElevenLabs models.
- Create `utils.py` for audio compression, file downloading, and cleanup functions.
